+++ apps/backend/src/modules/bookModule/application/commandHandlers/deleteUserBooksCommandHandler/deleteUserBooksCommandHandlerImpl.ts
@@ -2,6 +2,7 @@ import {
   type DeleteUserBooksCommandHandler,
   type DeleteUserBooksCommandHandlerPayload,
 } from './deleteUserBooksCommandHandler.js';
+import { ResourceNotFoundError } from '../../../../../common/errors/resourceNotFoundError.js';
 import { type LoggerService } from '../../../../../libs/logger/services/loggerService/loggerService.js';
 import { type UserBookRepository } from '../../../domain/repositories/userBookRepository/userBookRepository.js';
 
@@ -18,6 +19,19 @@ export class DeleteUserBooksCommandHandlerImpl implements DeleteUserBooksCommand
       message: `Deleting Users's books...`,
       ids: userBookIds,
     });
+
+    const existingUserBooks = await this.userBookRepository.findUserBooks({
+      ids: userBookIds,
+      page: 1,
+      pageSize: userBookIds.length,
+    });
+
+    if (existingUserBooks.length !== userBookIds.length) {
+      throw new ResourceNotFoundError({
+        resource: 'UserBook',
+        ids: userBookIds,
+      });
+    }
 
     await this.userBookRepository.deleteUserBooks({ ids: userBookIds });
